requirements.txt
line-bot-sdk
bs4
flask
pymongo
datetime
pandas
SnowNLP
emoji
pyshorteners
scipy
openai
google-generativeai